+++ docs/en/changes/changes.md
@@ -32,6 +32,7 @@
 * Add a component ID for Netty-http (ID=151).
 * Add a component ID for Fiber (ID=5021).
 * BanyanDBStorageClient: Add `define(Property property, PropertyStore.Strategy strategy)` API.
+* Support GraalVM native-image (Experimental).
 
 #### UI